+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Docker Service in YAML injects Linux X86_64 docker client
+  binary (Docker Service)
+
 ## [0.0.24] - 2019-12-21
 ### Added
 - `PIPELINES_PROJECT_PATH` parameter